url view done 1104

+++ GraudateProject/src/main/java/com/spring/graduateproject/url/service/IUrlService.java
@@ -4,5 +4,5 @@ public interface IUrlService {
 
 	
 	// url check
-	int urlCheck(String url);
+	double urlCheck(String url);
 }
